+++ changelog.md
@@ -1,5 +1,14 @@
 # ACEmulator Change Log
+### 2017-04-09
+[fantoms]
+* Added visual effect and sound broadcasting in action queue processing.
+* Added visual effect broadcasting to `Landblocks`.
+* Added `GameActionType` for `ApplySoundEfect` and `ApplyVisualEffect`.
+
 ### 2017-04-08
+[fantoms]
+* Addded sound broadcasting to `Landblocks`.
+
 [Zegeger]
 Added basic motion classes.
 Added current motion state to PhysicsData, which can now be sent in CreateObject messages.